Fix types in several declarations (patch by Konstantin Kuzov, 2:5019/40)

+++ fidoconf/stat.h
@@ -8,7 +8,7 @@
 /* stat record type */
 typedef enum { stNORM, stBAD, stDUPE, stOUT } st_type;
 
-HUSKYEXT void put_stat(s_area *echo, hs_addr *link, st_type type, long len);
+HUSKYEXT void put_stat(s_area *echo, hs_addr *link, st_type type, INT32 len);
 HUSKYEXT void upd_stat(char *file);
 
 #endif
